backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
ImageHash
opencv-python-headless
numpy
supabase
postgrest
python-dotenv
